% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRBEE_IPOD_SuSiE.R
\name{MRBEE_IPOD_SuSiE}
\alias{MRBEE_IPOD_SuSiE}
\title{Estimate Causal Effect with MRBEE and SuSiE}
\usage{
MRBEE_IPOD_SuSiE(
  by,
  bX,
  byse,
  bXse,
  Rxy,
  tauvec = seq(3, 30, 3),
  admm.rho = 3,
  Lvec = c(1:6),
  ebic.theta = 1,
  ebic.gamma = 2,
  susie.iter = 200,
  pip.thres = 0.3,
  max.iter = 50,
  max.eps = 1e-04,
  reliability.thres = 0.8,
  ridge.diff = 100,
  sampling.time = 100,
  sampling.iter = 10
)
}
\arguments{
\item{by}{A vector (n x 1) of the GWAS effect size of outcome.}

\item{bX}{A vector (n x 1) of the liner predictor of the transferred exposure's effect.}

\item{byse}{A vector (n x 1) of the GWAS effect size SE of outcome.}

\item{bXse}{A vector (n x 1) of the standard error of the liner predictor of the transferred exposure's effect.}

\item{Rxy}{A matrix (p+1 x p+1) of the correlation matrix of the p exposures and outcome. The first one should be the transferred linear predictor and last one should be the outcome.}

\item{tauvec}{The candidate vector of tuning parameters for the MCP penalty function. Default is \code{seq(3, 30, by=3)}.}

\item{admm.rho}{The tuning parameter in the nested ADMM algorithm. Default is \code{3}.}

\item{Lvec}{A vector of the number of single effects used in SuSiE. Default is \code{c(1:6)}.}

\item{ebic.gamma}{A scale of tuning parameter of horizontal pleiotropy in extended BIC. Default is \code{2}.}

\item{susie.iter}{A scale of the maximum number of iterations used in SuSiE. Default is \code{200}.}

\item{pip.thres}{A scale of PIP theshold for calibyating causality used in SuSiE. Default is \code{0.3}.}

\item{max.iter}{Maximum number of iterations for causal effect estimation. Default is \code{50}.}

\item{max.eps}{Tolerance for stopping criteria. Default is \code{1e-4}.}

\item{reliability.thres}{A scale of threshold for the minimum value of the reliability ratio. If the original reliability ratio is less than this threshold, only part of the estimation error is removed so that the working reliability ratio equals this threshold. Default is \code{0.8}.}

\item{ridge.diff}{A scale of parameter on the differences of causal effect estimate in one credible set. Defaults to \code{10}.}

\item{sampling.time}{A scale of number of subsampling in estimating the standard error. Default is \code{100}.}

\item{sampling.iter}{A scale of iteration in subsampling in estimating the standard error. Default is \code{10}.}

\item{ebic.delta}{A scale of tuning parameter of causal effect estimate in extended BIC. Default is \code{1}.}
}
\value{
A list containing the estimated causal effect, its covariance, and pleiotropy.
}
\description{
This function estimates the causal effect using a bias-correction estimating equation, considering potential pleiotropy and measurement errors, and using SuSiE to select the causal exposures.
}
