% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRcML_IPOD.R
\name{MRcML_IPOD}
\alias{MRcML_IPOD}
\title{Estimate Causal Effect with MRcML and SuSiE}
\usage{
MRcML_IPOD(
  by,
  bX,
  byse,
  bXse,
  Rxy,
  tau = 9,
  admm.rho = 3,
  max.iter = 10,
  max.eps = 1e-04,
  sampling.time = 100,
  sampling.iter = 10,
  theta.ini
)
}
\arguments{
\item{by}{A vector (n x 1) of the GWAS effect size of outcome.}

\item{bX}{A vector (n x 1) of the liner predictor of the transferred exposure's effect.}

\item{byse}{A vector (n x 1) of the GWAS effect size SE of outcome.}

\item{bXse}{A vector (n x 1) of the standard error of the liner predictor of the transferred exposure's effect.}

\item{Rxy}{A matrix (p+1 x p+1) of the correlation matrix of the p exposures and outcome. The first one should be the transferred linear predictor and last one should be the outcome.}

\item{tau}{A scale of tuning parameter in MCP. Defaults to \code{9}.}

\item{admm.rho}{A scale of tuning parameter in ADMM. Defaults to \code{3}.}

\item{max.iter}{Maximum number of iterations for causal effect estimation. Defaults to 10.}

\item{max.eps}{Tolerance for stopping criteria. Defaults to 1e-4.}

\item{sampling.time}{A scale of number of subsampling in estimating the standard error.}

\item{sampling.iter}{A scale of iteration in subsampling in estimating the standard error.}

\item{theta.ini}{An initial estimate of causal effect.}

\item{adjust.method}{Method for estimating q-value. Defaults to "Sidak".}
}
\value{
A list containing the estimated causal effect, its covariance, and pleiotropy.
}
\description{
This function estimates the causal effect using a bias-correction estimating equation, considering potential pleiotropy and measurement errors, and using SuSiE to select the causal exposures.
}
