% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRcML_TL.R
\name{MRcML_TL}
\alias{MRcML_TL}
\title{Estimate Non-Transferable Causal Effect with MRcML and SuSiE}
\usage{
MRcML_TL(
  by,
  bX,
  byse,
  bXse,
  Rxy,
  L = min(10, ncol(bX)),
  tau = 9,
  admm.rho = 3,
  susie.iter = 200,
  pip.thres = 0.2,
  max.iter = 10,
  max.eps = 1e-04,
  ridge.diff = 100,
  sampling.time = 100,
  sampling.iter = 10,
  theta.ini = F,
  theta.source,
  theta.source.cov,
  transfer.coef = 1
)
}
\arguments{
\item{by}{A vector (n x 1) of the GWAS effect size of outcome.}

\item{bX}{A vector (n x 1) of the liner predictor of the transferred exposure's effect.}

\item{byse}{A vector (n x 1) of the GWAS effect size SE of outcome.}

\item{bXse}{A vector (n x 1) of the standard error of the liner predictor of the transferred exposure's effect.}

\item{Rxy}{A matrix (p+1 x p+1) of the correlation matrix of the p exposures and outcome. The first one should be the transferred linear predictor and last one should be the outcome.}

\item{L}{A scale of the number of single effects used in SuSiE. Defaults to \code{5}.}

\item{tau}{A scale of tuning parameter in MCP. Defaults to \code{9}.}

\item{admm.rho}{A scale of tuning parameter in ADMM. Defaults to \code{3}.}

\item{susie.iter}{A scale of the maximum number of iterations used in SuSiE. Defaults to \code{200}.}

\item{pip.thres}{A scale of PIP threshold for calibyating causality used in SuSiE.}

\item{max.iter}{Maximum number of iterations for causal effect estimation. Defaults to \code{10}.}

\item{max.eps}{Tolerance for stopping criteria. Defaults to \code{1e-4}.}

\item{ridge.diff}{A scale of parameter on the differences of causal effect estimate in one credible set. Defaults to \code{10}.}

\item{sampling.time}{A scale of number of subsampling in estimating the standard error. Defaults to \code{100}.}

\item{sampling.iter}{A scale of iteration in subsampling in estimating the standard error. Defaults to \code{10}.}

\item{theta.ini}{An initial estimate of causal effect.}

\item{theta.source}{A vector (p x 1) of the causal effect estimate learning from the source data.}

\item{theta.source.cov}{A matrix (p x p) of the covariance matrix of the causal effect estimate learning from the source data.}

\item{transfer.coef}{A scale of transfer.coef of theta.source to theta.target. Default is \code{1}.}
}
\value{
A list containing the estimated causal effect, its covariance, and pleiotropy.
}
\description{
This function estimates the non-transferable causal effect using a conditional score method, considering potential pleiotropy and measurement errors, and using SuSiE to select the non-transferable causal effect.
}
