% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TR_MRBEE_IMRP_UV.R
\name{TR_MRBEE_IMRP_UV}
\alias{TR_MRBEE_IMRP_UV}
\title{Estimate Non-Transferable Causal Effect with Univariable MRBEE and SuSiE}
\usage{
TR_MRBEE_IMRP_UV(
  by,
  bx,
  byse,
  bxse,
  Rxy,
  theta.source,
  susie.iter = 100,
  pip.thres = 0.3,
  max.iter = 30,
  max.eps = 1e-04,
  pv.thres = 0.05,
  var.est = "variance",
  FDR = T,
  adjust.method = "Sidak"
)
}
\arguments{
\item{by}{Vector of GWAS effect sizes for the outcome (n x 1).}

\item{bx}{Vector of GWAS effect sizes for the exposure (n x 1).}

\item{byse}{Vector of standard errors (SE) for the GWAS effect sizes of the outcome (n x 1).}

\item{bxse}{Vector of SE for the GWAS effect sizes of the exposure (n x 1).}

\item{Rxy}{Correlation matrix (p+1 x p+1) of the exposures and outcome, with the outcome being the last.}

\item{theta.source}{A vector (p x 1) of the causal effect estimate learning from the source data.}

\item{susie.iter}{A scale of the maximum number of iterations used in SuSiE.}

\item{pip.thres}{A scale of PIP theshold for calibyating causality used in SuSiE.}

\item{max.iter}{Maximum number of iterations for the estimation process. Defaults to 30.}

\item{max.eps}{Tolerance level for convergence. Defaults to 1e-4.}

\item{pv.thres}{P-value threshold for pleiotropy detection. Defaults to 0.05.}

\item{var.est}{Method for estimating the standard error in the pleiotropy test. Can be "robust", "variance", or "ordinal".}

\item{FDR}{Logical indicating whether to apply False Discovery Rate (FDR) correction. Defaults to TRUE.}

\item{adjust.method}{Method for estimating q-values, defaults to "Sidak".}
}
\value{
A list containing the estimated causal effect, its covariance, and pleiotropy
}
\description{
This function performs univariable Mendelian Randomization (MR) analysis using
the MRBEE method.
}
